Extract cutpoints from a UD model in a tidy format.

This function takes a model of the democracy scores and extracts the discrimination parameters, score cutpoints, and standard errors for all the variables involved, putting these into a tidy data frame.

cutpoints(model, type = "score")

Arguments

model
A mirt SingleGroupClass-class model of the democracy scores.

type
A string specifying the cutpoint type. Can be (an abbreviation of) "score" (for score cutpoints) or "discrimination" (for discrimination parameters). Default is "score."

Value

A tibble with either score cutpoints for each variable used to construct the latent scores in terms of the latent variable (the default), or discrimination parameters for each variable used to construct the index. For the score cutpoints (type = 'score'), the columns estimate, pct975, and pct025 report the IRT parametrization of the model estimates, a normalized measure in the same scale as the latent variable.

Examples

# Replicate the official UDS 2011 release and calculate its cutpoints
replication_2011_model <- democracy_model(dplyr::matches("pmm"), verbose = FALSE)
cutpoints(replication_2011_model)

#> # A tibble: 82 x 6
#>    variable   estimate    pct025  pct975      se num_obs
#>    <chr>         <dbl>     <dbl>   <dbl>   <dbl>   <int>
#>  1 pmm_arat   -1.43    -1.44     -1.42   0.00473    3873
#>  2 pmm_arat   -1.02    -1.01     -1.02   0.00136    3873
#>  3 pmm_arat   -0.427   -0.402    -0.448  0.0111     3873
#>  4 pmm_arat   -0.0419  -0.000439 -0.0792 0.0190     3873
#>  5 pmm_arat    0.421    0.492     0.357  0.0326     3873
#>  6 pmm_arat    1.42     1.58      1.28   0.0719     3873
#>  7 pmm_blm    -0.00248  0.0907   -0.0445 0.0214      275
#>  8 pmm_blm     0.474    1.04      0.220  0.130       275
#>  9 pmm_bollen -1.53    -1.55     -1.50   0.0109      510
#> 10 pmm_bollen -1.08    -1.08     -1.07   0.00183     510
#> # ... with 72 more rows