Extract UD scores from a UD model

This function is a simple wrapper for fscores(model, full.scores = TRUE, full.scores.SE = TRUE, ...) that returns scores in a tidy data frame instead of a matrix. More fine-grained control can be achieved by using fscores directly.

democracy_scores(model, ...)

Arguments

model

a SingleGroupClass-class model produced by democracy_model.

...

Other parameters passed on to fscores.

Value

A data frame with latent variable democracy scores (posterior means) for all country-years in the data, with standard errors and 95 intervals.

Examples

# Replicate the official UDS scores (2011 release)
library(reshape2)
library(dplyr)
data <- prepare_data(democracy)
data <- melt(data, measure.vars = names(data)[grep("pmm",names(data))], na.rm = TRUE)
data <- data %>% group_by(country_name,year) %>% mutate(num_measures = n())
data <- dcast(data, ... ~ variable)
data <- data %>% arrange(country_name,year)
replication_2011_model <- democracy_model(data,
                           columns = names(data)[grep("pmm",names(data))])

#> Iteration: 1, Log-Lik: -76709.696, Max-Change: 1.69565Iteration: 2, Log-Lik: -67424.210, Max-Change: 1.81540Iteration: 3, Log-Lik: -62044.869, Max-Change: 1.90586Iteration: 4, Log-Lik: -58462.239, Max-Change: 2.11255Iteration: 5, Log-Lik: -57044.823, Max-Change: 1.99305Iteration: 6, Log-Lik: -56419.469, Max-Change: 0.75299Iteration: 7, Log-Lik: -56213.275, Max-Change: 0.59372Iteration: 8, Log-Lik: -56084.377, Max-Change: 0.65020Iteration: 9, Log-Lik: -56002.673, Max-Change: 0.34590Iteration: 10, Log-Lik: -55946.122, Max-Change: 0.40502Iteration: 11, Log-Lik: -55911.979, Max-Change: 0.30872Iteration: 12, Log-Lik: -55882.074, Max-Change: 0.31877Iteration: 13, Log-Lik: -55859.401, Max-Change: 0.23352Iteration: 14, Log-Lik: -55839.372, Max-Change: 0.27760Iteration: 15, Log-Lik: -55820.964, Max-Change: 0.26610Iteration: 16, Log-Lik: -55805.350, Max-Change: 0.17899Iteration: 17, Log-Lik: -55793.394, Max-Change: 0.19857Iteration: 18, Log-Lik: -55783.677, Max-Change: 0.21322Iteration: 19, Log-Lik: -55774.554, Max-Change: 0.17610Iteration: 20, Log-Lik: -55766.512, Max-Change: 0.18911Iteration: 21, Log-Lik: -55759.718, Max-Change: 0.15640Iteration: 22, Log-Lik: -55754.031, Max-Change: 0.17177Iteration: 23, Log-Lik: -55748.773, Max-Change: 0.14150Iteration: 24, Log-Lik: -55743.646, Max-Change: 0.11852Iteration: 25, Log-Lik: -55739.065, Max-Change: 0.12276Iteration: 26, Log-Lik: -55735.703, Max-Change: 0.14926Iteration: 27, Log-Lik: -55732.614, Max-Change: 0.10478Iteration: 28, Log-Lik: -55727.620, Max-Change: 0.09749Iteration: 29, Log-Lik: -55725.048, Max-Change: 0.10135Iteration: 30, Log-Lik: -55722.638, Max-Change: 0.09671Iteration: 31, Log-Lik: -55720.283, Max-Change: 0.06820Iteration: 32, Log-Lik: -55718.617, Max-Change: 0.07155Iteration: 33, Log-Lik: -55717.245, Max-Change: 0.10022Iteration: 34, Log-Lik: -55714.900, Max-Change: 0.05632Iteration: 35, Log-Lik: -55713.615, Max-Change: 0.08139Iteration: 36, Log-Lik: -55712.642, Max-Change: 0.36362Iteration: 37, Log-Lik: -55711.006, Max-Change: 0.05699Iteration: 38, Log-Lik: -55709.395, Max-Change: 0.03548Iteration: 39, Log-Lik: -55708.714, Max-Change: 0.47823Iteration: 40, Log-Lik: -55707.055, Max-Change: 0.03267Iteration: 41, Log-Lik: -55706.615, Max-Change: 0.53661Iteration: 42, Log-Lik: -55705.342, Max-Change: 0.03394Iteration: 43, Log-Lik: -55705.270, Max-Change: 0.02693Iteration: 44, Log-Lik: -55704.793, Max-Change: 0.29289Iteration: 45, Log-Lik: -55704.052, Max-Change: 0.08215Iteration: 46, Log-Lik: -55703.821, Max-Change: 0.34977Iteration: 47, Log-Lik: -55703.158, Max-Change: 0.09499Iteration: 48, Log-Lik: -55702.775, Max-Change: 0.21498Iteration: 49, Log-Lik: -55702.131, Max-Change: 0.05538Iteration: 50, Log-Lik: -55701.833, Max-Change: 0.19621Iteration: 51, Log-Lik: -55701.480, Max-Change: 0.08780Iteration: 52, Log-Lik: -55701.249, Max-Change: 0.01101Iteration: 53, Log-Lik: -55701.053, Max-Change: 0.12021Iteration: 54, Log-Lik: -55700.817, Max-Change: 0.01783Iteration: 55, Log-Lik: -55700.776, Max-Change: 0.01126Iteration: 56, Log-Lik: -55700.617, Max-Change: 0.12523Iteration: 57, Log-Lik: -55700.394, Max-Change: 0.01934Iteration: 58, Log-Lik: -55700.353, Max-Change: 0.07486Iteration: 59, Log-Lik: -55700.174, Max-Change: 0.02268Iteration: 60, Log-Lik: -55700.024, Max-Change: 0.02004Iteration: 61, Log-Lik: -55699.618, Max-Change: 0.00557Iteration: 62, Log-Lik: -55699.501, Max-Change: 0.00532Iteration: 63, Log-Lik: -55699.416, Max-Change: 0.03192Iteration: 64, Log-Lik: -55699.279, Max-Change: 0.00945Iteration: 65, Log-Lik: -55699.197, Max-Change: 0.01170Iteration: 66, Log-Lik: -55699.118, Max-Change: 0.00867Iteration: 67, Log-Lik: -55698.901, Max-Change: 0.07198Iteration: 68, Log-Lik: -55698.769, Max-Change: 0.00819Iteration: 69, Log-Lik: -55698.709, Max-Change: 0.01809Iteration: 70, Log-Lik: -55698.618, Max-Change: 0.00525Iteration: 71, Log-Lik: -55698.572, Max-Change: 0.00465Iteration: 72, Log-Lik: -55698.543, Max-Change: 0.00505Iteration: 73, Log-Lik: -55698.494, Max-Change: 0.02106Iteration: 74, Log-Lik: -55698.424, Max-Change: 0.00296Iteration: 75, Log-Lik: -55698.398, Max-Change: 0.00542Iteration: 76, Log-Lik: -55698.330, Max-Change: 0.02606Iteration: 77, Log-Lik: -55698.269, Max-Change: 0.00494Iteration: 78, Log-Lik: -55698.244, Max-Change: 0.00302Iteration: 79, Log-Lik: -55698.235, Max-Change: 0.00383Iteration: 80, Log-Lik: -55698.212, Max-Change: 0.00301Iteration: 81, Log-Lik: -55698.198, Max-Change: 0.02207Iteration: 82, Log-Lik: -55698.141, Max-Change: 0.00341Iteration: 83, Log-Lik: -55698.132, Max-Change: 0.00333Iteration: 84, Log-Lik: -55698.108, Max-Change: 0.00637Iteration: 85, Log-Lik: -55698.062, Max-Change: 0.00373Iteration: 86, Log-Lik: -55698.048, Max-Change: 0.00268Iteration: 87, Log-Lik: -55698.037, Max-Change: 0.00298Iteration: 88, Log-Lik: -55698.006, Max-Change: 0.00165Iteration: 89, Log-Lik: -55697.987, Max-Change: 0.00055Iteration: 90, Log-Lik: -55697.985, Max-Change: 0.02519Iteration: 91, Log-Lik: -55697.942, Max-Change: 0.00176Iteration: 92, Log-Lik: -55697.934, Max-Change: 0.00338Iteration: 93, Log-Lik: -55697.924, Max-Change: 0.00274Iteration: 94, Log-Lik: -55697.910, Max-Change: 0.00175Iteration: 95, Log-Lik: -55697.906, Max-Change: 0.00130Iteration: 96, Log-Lik: -55697.904, Max-Change: 0.00076Iteration: 97, Log-Lik: -55697.902, Max-Change: 0.00085Iteration: 98, Log-Lik: -55697.900, Max-Change: 0.00054Iteration: 99, Log-Lik: -55697.898, Max-Change: 0.00045Iteration: 100, Log-Lik: -55697.898, Max-Change: 0.00542Iteration: 101, Log-Lik: -55697.886, Max-Change: 0.00063Iteration: 102, Log-Lik: -55697.884, Max-Change: 0.00024Iteration: 103, Log-Lik: -55697.883, Max-Change: 0.00059Iteration: 104, Log-Lik: -55697.882, Max-Change: 0.00043Iteration: 105, Log-Lik: -55697.882, Max-Change: 0.00042Iteration: 106, Log-Lik: -55697.881, Max-Change: 0.00395Iteration: 107, Log-Lik: -55697.872, Max-Change: 0.00042Iteration: 108, Log-Lik: -55697.871, Max-Change: 0.00029Iteration: 109, Log-Lik: -55697.871, Max-Change: 0.00600Iteration: 110, Log-Lik: -55697.859, Max-Change: 0.00156Iteration: 111, Log-Lik: -55697.857, Max-Change: 0.00093Iteration: 112, Log-Lik: -55697.854, Max-Change: 0.00042Iteration: 113, Log-Lik: -55697.854, Max-Change: 0.00030Iteration: 114, Log-Lik: -55697.854, Max-Change: 0.00019Iteration: 115, Log-Lik: -55697.853, Max-Change: 0.01363Iteration: 116, Log-Lik: -55697.833, Max-Change: 0.00097Iteration: 117, Log-Lik: -55697.830, Max-Change: 0.00057Iteration: 118, Log-Lik: -55697.828, Max-Change: 0.00027Iteration: 119, Log-Lik: -55697.828, Max-Change: 0.00022Iteration: 120, Log-Lik: -55697.827, Max-Change: 0.00017Iteration: 121, Log-Lik: -55697.827, Max-Change: 0.00133Iteration: 122, Log-Lik: -55697.823, Max-Change: 0.00058Iteration: 123, Log-Lik: -55697.822, Max-Change: 0.00022Iteration: 124, Log-Lik: -55697.821, Max-Change: 0.00046Iteration: 125, Log-Lik: -55697.821, Max-Change: 0.00064Iteration: 126, Log-Lik: -55697.820, Max-Change: 0.00044Iteration: 127, Log-Lik: -55697.819, Max-Change: 0.00047Iteration: 128, Log-Lik: -55697.819, Max-Change: 0.00027Iteration: 129, Log-Lik: -55697.818, Max-Change: 0.00026Iteration: 130, Log-Lik: -55697.818, Max-Change: 0.00236Iteration: 131, Log-Lik: -55697.812, Max-Change: 0.00022Iteration: 132, Log-Lik: -55697.811, Max-Change: 0.00017Iteration: 133, Log-Lik: -55697.811, Max-Change: 0.00073Iteration: 134, Log-Lik: -55697.809, Max-Change: 0.00017Iteration: 135, Log-Lik: -55697.809, Max-Change: 0.00014Iteration: 136, Log-Lik: -55697.809, Max-Change: 0.00647Iteration: 137, Log-Lik: -55697.803, Max-Change: 0.00345Iteration: 138, Log-Lik: -55697.798, Max-Change: 0.00073Iteration: 139, Log-Lik: -55697.797, Max-Change: 0.00069Iteration: 140, Log-Lik: -55697.795, Max-Change: 0.00016Iteration: 141, Log-Lik: -55697.795, Max-Change: 0.00012Iteration: 142, Log-Lik: -55697.794, Max-Change: 0.00095Iteration: 143, Log-Lik: -55697.792, Max-Change: 0.00012Iteration: 144, Log-Lik: -55697.792, Max-Change: 0.00008
#> 
#> Calculating information matrix...

replication_2011_scores <- democracy_scores(replication_2011_model)
replication_2011_scores <- bind_cols(data %>%
                                       select(country_name,
                                       GWn,
                                       year,
                                       in_system,
                                       num_measures),
                                       replication_2011_scores)
head(replication_2011_scores)

#>   country_name GWn year in_system num_measures        z1     se.z1     pct975
#> 1  Afghanistan 700 1946      TRUE            4 -1.885820 0.4863174 -0.9326377
#> 2  Afghanistan 700 1947      TRUE            4 -1.885820 0.4863174 -0.9326377
#> 3  Afghanistan 700 1948      TRUE            5 -1.573819 0.3234092 -0.9399366
#> 4  Afghanistan 700 1949      TRUE            5 -1.573819 0.3234092 -0.9399366
#> 5  Afghanistan 700 1950      TRUE            6 -1.447628 0.2294827 -0.9978421
#> 6  Afghanistan 700 1951      TRUE            5 -1.573819 0.3234092 -0.9399366
#>      pct025
#> 1 -2.839002
#> 2 -2.839002
#> 3 -2.207701
#> 4 -2.207701
#> 5 -1.897414
#> 6 -2.207701