Match generated scores to the level of an official UD release

Match generated scores to the level of an official UD release

match_to_uds(data, release = 2014)

Arguments

data

A dataset containing an extension of the UD scores, from the output of democracy_scores. It must contain a country_name column, a year column, z1 column, a pct025 column, and a pct975 column.

release

The year of a Unified Democracy Scores official release (see the available releases in their data page at http://www.unified-democracy-scores.org/uds.html and http://www.unified-democracy-scores.org/archive.html). Default is 2014.

Value

The dataset with three columns appended: adj.z1 (the mean democracy score, adjusted to match the UD release), adj.pct025 and adj.pct975 (the adjusted confidence bounds).

Examples

library(dplyr)
indexes <- c("arat_pmm","blm","bmr_democracy","bollen_pmm",
                            "doorenspleet","eiu","freedomhouse",
                            "gwf","hadenius_pmm","lied","mainwaring",
                            "munck_pmm","pacl","polity2","polyarchy_pmm","prc",
                            "svolik","ulfelder","vanhanen_democratization",
                            "wahman_teorell_hadenius")
data <- prepare_democracy(indexes)
extended_model <- democracy_model(data,indexes, verbose=TRUE)

#> Iteration: 1, Log-Lik: -200107.540, Max-Change: 2.85757Iteration: 2, Log-Lik: -162758.584, Max-Change: 1.33910Iteration: 3, Log-Lik: -149529.031, Max-Change: 1.93529Iteration: 4, Log-Lik: -141676.598, Max-Change: 2.13551Iteration: 5, Log-Lik: -135043.660, Max-Change: 1.26942Iteration: 6, Log-Lik: -132496.552, Max-Change: 1.16428Iteration: 7, Log-Lik: -130610.893, Max-Change: 0.81043Iteration: 8, Log-Lik: -129926.875, Max-Change: 1.21628Iteration: 9, Log-Lik: -129261.697, Max-Change: 0.42356Iteration: 10, Log-Lik: -128848.091, Max-Change: 0.32146Iteration: 11, Log-Lik: -128465.304, Max-Change: 0.30539Iteration: 12, Log-Lik: -128145.677, Max-Change: 0.33638Iteration: 13, Log-Lik: -127859.510, Max-Change: 0.33051Iteration: 14, Log-Lik: -127644.267, Max-Change: 0.25711Iteration: 15, Log-Lik: -127450.770, Max-Change: 0.29728Iteration: 16, Log-Lik: -127225.323, Max-Change: 0.16220Iteration: 17, Log-Lik: -127053.489, Max-Change: 0.17796Iteration: 18, Log-Lik: -126932.899, Max-Change: 0.18044Iteration: 19, Log-Lik: -126830.823, Max-Change: 0.19177Iteration: 20, Log-Lik: -126747.942, Max-Change: 0.18972Iteration: 21, Log-Lik: -126664.672, Max-Change: 0.19765Iteration: 22, Log-Lik: -126599.968, Max-Change: 0.18962Iteration: 23, Log-Lik: -126540.821, Max-Change: 0.18143Iteration: 24, Log-Lik: -126508.362, Max-Change: 0.27994Iteration: 25, Log-Lik: -126458.167, Max-Change: 0.16007Iteration: 26, Log-Lik: -126416.628, Max-Change: 0.15298Iteration: 27, Log-Lik: -126381.647, Max-Change: 0.14834Iteration: 28, Log-Lik: -126354.878, Max-Change: 0.15359Iteration: 29, Log-Lik: -126328.169, Max-Change: 0.14601Iteration: 30, Log-Lik: -126306.340, Max-Change: 0.16275Iteration: 31, Log-Lik: -126286.281, Max-Change: 0.14742Iteration: 32, Log-Lik: -126263.231, Max-Change: 0.23568Iteration: 33, Log-Lik: -126244.140, Max-Change: 0.12107Iteration: 34, Log-Lik: -126227.642, Max-Change: 0.13396Iteration: 35, Log-Lik: -126212.867, Max-Change: 0.10527Iteration: 36, Log-Lik: -126201.514, Max-Change: 0.08871Iteration: 37, Log-Lik: -126187.152, Max-Change: 0.15287Iteration: 38, Log-Lik: -126175.981, Max-Change: 0.09165Iteration: 39, Log-Lik: -126165.529, Max-Change: 0.10636Iteration: 40, Log-Lik: -126155.427, Max-Change: 0.06706Iteration: 41, Log-Lik: -126145.306, Max-Change: 0.07131Iteration: 42, Log-Lik: -126137.439, Max-Change: 0.07008Iteration: 43, Log-Lik: -126129.771, Max-Change: 0.12581Iteration: 44, Log-Lik: -126121.605, Max-Change: 0.07682Iteration: 45, Log-Lik: -126115.049, Max-Change: 0.07006Iteration: 46, Log-Lik: -126107.978, Max-Change: 0.13103Iteration: 47, Log-Lik: -126102.936, Max-Change: 0.05237Iteration: 48, Log-Lik: -126098.106, Max-Change: 0.10133Iteration: 49, Log-Lik: -126092.819, Max-Change: 0.05323Iteration: 50, Log-Lik: -126088.525, Max-Change: 0.04108Iteration: 51, Log-Lik: -126084.805, Max-Change: 0.07147Iteration: 52, Log-Lik: -126080.339, Max-Change: 0.08202Iteration: 53, Log-Lik: -126074.620, Max-Change: 0.03183Iteration: 54, Log-Lik: -126071.571, Max-Change: 0.03113Iteration: 55, Log-Lik: -126067.844, Max-Change: 0.03805Iteration: 56, Log-Lik: -126065.818, Max-Change: 0.03268Iteration: 57, Log-Lik: -126063.719, Max-Change: 0.03778Iteration: 58, Log-Lik: -126063.144, Max-Change: 0.37144Iteration: 59, Log-Lik: -126053.805, Max-Change: 0.03134Iteration: 60, Log-Lik: -126051.977, Max-Change: 0.01945Iteration: 61, Log-Lik: -126050.205, Max-Change: 0.01151Iteration: 62, Log-Lik: -126049.295, Max-Change: 0.09830Iteration: 63, Log-Lik: -126048.283, Max-Change: 0.05312Iteration: 64, Log-Lik: -126047.581, Max-Change: 0.40199Iteration: 65, Log-Lik: -126046.226, Max-Change: 0.02641Iteration: 66, Log-Lik: -126045.072, Max-Change: 0.01285Iteration: 67, Log-Lik: -126044.248, Max-Change: 0.01037Iteration: 68, Log-Lik: -126043.729, Max-Change: 0.23336Iteration: 69, Log-Lik: -126043.142, Max-Change: 0.04450Iteration: 70, Log-Lik: -126042.973, Max-Change: 0.02286Iteration: 71, Log-Lik: -126042.406, Max-Change: 0.01076Iteration: 72, Log-Lik: -126041.949, Max-Change: 0.00895Iteration: 73, Log-Lik: -126040.045, Max-Change: 0.01653Iteration: 74, Log-Lik: -126039.631, Max-Change: 0.00923Iteration: 75, Log-Lik: -126039.429, Max-Change: 0.07338Iteration: 76, Log-Lik: -126038.975, Max-Change: 0.00709Iteration: 77, Log-Lik: -126038.755, Max-Change: 0.01008Iteration: 78, Log-Lik: -126038.535, Max-Change: 0.00580Iteration: 79, Log-Lik: -126038.268, Max-Change: 0.00652Iteration: 80, Log-Lik: -126038.109, Max-Change: 0.36770Iteration: 81, Log-Lik: -126037.643, Max-Change: 0.02914Iteration: 82, Log-Lik: -126037.615, Max-Change: 0.02803Iteration: 83, Log-Lik: -126037.375, Max-Change: 0.00657Iteration: 84, Log-Lik: -126037.225, Max-Change: 0.00562Iteration: 85, Log-Lik: -126036.991, Max-Change: 0.01635Iteration: 86, Log-Lik: -126036.522, Max-Change: 0.00463Iteration: 87, Log-Lik: -126036.414, Max-Change: 0.00394Iteration: 88, Log-Lik: -126036.355, Max-Change: 0.02676Iteration: 89, Log-Lik: -126036.185, Max-Change: 0.00574Iteration: 90, Log-Lik: -126036.085, Max-Change: 0.00408Iteration: 91, Log-Lik: -126035.948, Max-Change: 0.00324Iteration: 92, Log-Lik: -126035.886, Max-Change: 0.00318Iteration: 93, Log-Lik: -126035.823, Max-Change: 0.00615Iteration: 94, Log-Lik: -126035.649, Max-Change: 0.00307Iteration: 95, Log-Lik: -126035.583, Max-Change: 0.00384Iteration: 96, Log-Lik: -126035.530, Max-Change: 0.02957Iteration: 97, Log-Lik: -126035.427, Max-Change: 0.00109Iteration: 98, Log-Lik: -126035.424, Max-Change: 0.00249Iteration: 99, Log-Lik: -126035.395, Max-Change: 0.00124Iteration: 100, Log-Lik: -126035.386, Max-Change: 0.00355Iteration: 101, Log-Lik: -126035.341, Max-Change: 0.00235Iteration: 102, Log-Lik: -126035.310, Max-Change: 0.00267Iteration: 103, Log-Lik: -126035.295, Max-Change: 0.00335Iteration: 104, Log-Lik: -126035.170, Max-Change: 0.00014Iteration: 105, Log-Lik: -126035.168, Max-Change: 0.00284Iteration: 106, Log-Lik: -126035.139, Max-Change: 0.00156Iteration: 107, Log-Lik: -126035.129, Max-Change: 0.00017Iteration: 108, Log-Lik: -126035.128, Max-Change: 0.00032Iteration: 109, Log-Lik: -126035.127, Max-Change: 0.00265Iteration: 110, Log-Lik: -126035.098, Max-Change: 0.00066Iteration: 111, Log-Lik: -126035.094, Max-Change: 0.00054Iteration: 112, Log-Lik: -126035.090, Max-Change: 0.00015Iteration: 113, Log-Lik: -126035.089, Max-Change: 0.00200Iteration: 114, Log-Lik: -126035.068, Max-Change: 0.00014Iteration: 115, Log-Lik: -126035.068, Max-Change: 0.00014Iteration: 116, Log-Lik: -126035.067, Max-Change: 0.00099Iteration: 117, Log-Lik: -126035.062, Max-Change: 0.00247Iteration: 118, Log-Lik: -126035.044, Max-Change: 0.00071Iteration: 119, Log-Lik: -126035.030, Max-Change: 0.00012Iteration: 120, Log-Lik: -126035.029, Max-Change: 0.00077Iteration: 121, Log-Lik: -126035.025, Max-Change: 0.00026Iteration: 122, Log-Lik: -126035.024, Max-Change: 0.00022Iteration: 123, Log-Lik: -126035.023, Max-Change: 0.00029Iteration: 124, Log-Lik: -126035.022, Max-Change: 0.00030Iteration: 125, Log-Lik: -126035.021, Max-Change: 0.00008
#> 
#> Calculating information matrix...

extended_scores <- democracy_scores(extended_model)
extended_scores <- bind_cols(data,extended_scores)
# Matches the calculated scores to the level of the 2014 release of the UD scores
extended_scores <- match_to_uds(extended_scores)