Prepares selected indexes in the democracy dataset for use in a UD model

Prepares selected indexes in the democracy dataset for use in a UD model

prepare_democracy(indexes)

Arguments

indexes

a set of measures of democracy in the democracy dataset. They must be names of the columns of the democracy dataset - otherwise the function throws an error.

Value

A dataset ready for use with democracy_model

Examples

prepare_democracy(names(democracy)[grep("pmm",names(democracy))])

#>                                 country_name GWn GWc cown polity_ccode year
#> 1                                Afghanistan 700 AFG  700          159 1946
#> 2                                Afghanistan 700 AFG  700          159 1947
#> 3                                Afghanistan 700 AFG  700          159 1948
#> 4                                Afghanistan 700 AFG  700          159 1949
#> 5                                Afghanistan 700 AFG  700          159 1950
#> 6                                Afghanistan 700 AFG  700          159 1951
#> 7                                Afghanistan 700 AFG  700          159 1952
#> 8                                Afghanistan 700 AFG  700          159 1953
#> 9                                Afghanistan 700 AFG  700          159 1954
#> 10                               Afghanistan 700 AFG  700          159 1955
#> 11                               Afghanistan 700 AFG  700          159 1956
#> 12                               Afghanistan 700 AFG  700          159 1957
#> 13                               Afghanistan 700 AFG  700          159 1958
#> 14                               Afghanistan 700 AFG  700          159 1959
#> 15                               Afghanistan 700 AFG  700          159 1960
#> 16                               Afghanistan 700 AFG  700          159 1961
#> 17                               Afghanistan 700 AFG  700          159 1962
#> 18                               Afghanistan 700 AFG  700          159 1963
#> 19                               Afghanistan 700 AFG  700          159 1964
#> 20                               Afghanistan 700 AFG  700          159 1965
#> 21                               Afghanistan 700 AFG  700          159 1966
#> 22                               Afghanistan 700 AFG  700          159 1967
#> 23                               Afghanistan 700 AFG  700          159 1968
#> 24                               Afghanistan 700 AFG  700          159 1969
#> 25                               Afghanistan 700 AFG  700          159 1970
#> 26                               Afghanistan 700 AFG  700          159 1971
#> 27                               Afghanistan 700 AFG  700          159 1972
#> 28                               Afghanistan 700 AFG  700          159 1973
#> 29                               Afghanistan 700 AFG  700          159 1974
#> 30                               Afghanistan 700 AFG  700          159 1975
#> 31                               Afghanistan 700 AFG  700          159 1976
#> 32                               Afghanistan 700 AFG  700          159 1977
#> 33                               Afghanistan 700 AFG  700          159 1978
#> 34                               Afghanistan 700 AFG  700          159 1979
#> 35                               Afghanistan 700 AFG  700          159 1980
#> 36                               Afghanistan 700 AFG  700          159 1981
#> 37                               Afghanistan 700 AFG  700          159 1982
#>      GW_startdate GW_enddate                    region continent microstate
#> 1      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 2      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 3      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 4      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 5      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 6      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 7      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 8      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 9      1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 10     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 11     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 12     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 13     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 14     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 15     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 16     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 17     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 18     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 19     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 20     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 21     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 22     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 23     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 24     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 25     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 26     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 27     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 28     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 29     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 30     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 31     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 32     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 33     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 34     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 35     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 36     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#> 37     1919-05-01       <NA>             Southern Asia      Asia      FALSE
#>             lat         lon in_system in_cow arat_pmm blm_pmm bollen_pmm
#> 1     33.939110   67.709953      TRUE   TRUE       NA      NA         NA
#> 2     33.939110   67.709953      TRUE   TRUE       NA      NA         NA
#> 3     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 4     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 5     33.939110   67.709953      TRUE   TRUE        2      NA          2
#> 6     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 7     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 8     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 9     33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 10    33.939110   67.709953      TRUE   TRUE        2      NA          2
#> 11    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 12    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 13    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 14    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 15    33.939110   67.709953      TRUE   TRUE        2      NA          2
#> 16    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 17    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 18    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 19    33.939110   67.709953      TRUE   TRUE        1      NA         NA
#> 20    33.939110   67.709953      TRUE   TRUE        3      NA          3
#> 21    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 22    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 23    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 24    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 25    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 26    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 27    33.939110   67.709953      TRUE   TRUE        3      NA         NA
#> 28    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 29    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 30    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 31    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 32    33.939110   67.709953      TRUE   TRUE        2      NA         NA
#> 33    33.939110   67.709953      TRUE   TRUE        1      NA         NA
#> 34    33.939110   67.709953      TRUE   TRUE        1      NA         NA
#> 35    33.939110   67.709953      TRUE   TRUE        1      NA          1
#> 36    33.939110   67.709953      TRUE   TRUE        1      NA         NA
#> 37    33.939110   67.709953      TRUE   TRUE        1      NA         NA
#>      freedomhouse_pmm hadenius_pmm mainwaring_pmm munck_pmm pacl_pmm polity_pmm
#> 1                  NA           NA             NA        NA        1          1
#> 2                  NA           NA             NA        NA        1          1
#> 3                  NA           NA             NA        NA        1          1
#> 4                  NA           NA             NA        NA        1          1
#> 5                  NA           NA             NA        NA        1          1
#> 6                  NA           NA             NA        NA        1          1
#> 7                  NA           NA             NA        NA        1          1
#> 8                  NA           NA             NA        NA        1          1
#> 9                  NA           NA             NA        NA        1          1
#> 10                 NA           NA             NA        NA        1          1
#> 11                 NA           NA             NA        NA        1          1
#> 12                 NA           NA             NA        NA        1          1
#> 13                 NA           NA             NA        NA        1          1
#> 14                 NA           NA             NA        NA        1          1
#> 15                 NA           NA             NA        NA        1          1
#> 16                 NA           NA             NA        NA        1          1
#> 17                 NA           NA             NA        NA        1          1
#> 18                 NA           NA             NA        NA        1          1
#> 19                 NA           NA             NA        NA        1          4
#> 20                 NA           NA             NA        NA        1          4
#> 21                 NA           NA             NA        NA        1          4
#> 22                 NA           NA             NA        NA        1          4
#> 23                 NA           NA             NA        NA        1          4
#> 24                 NA           NA             NA        NA        1          4
#> 25                 NA           NA             NA        NA        1          4
#> 26                 NA           NA             NA        NA        1          4
#> 27                  6           NA             NA        NA        1          4
#> 28                  2           NA             NA        NA        1          4
#> 29                  2           NA             NA        NA        1          4
#> 30                  2           NA             NA        NA        1          4
#> 31                  2           NA             NA        NA        1          4
#> 32                  3           NA             NA        NA        1          4
#> 33                  1           NA             NA        NA        1         11
#> 34                  1           NA             NA        NA        1         NA
#> 35                  1           NA             NA        NA        1         NA
#> 36                  1           NA             NA        NA        1         NA
#> 37                  1           NA             NA        NA        1         NA
#>      polyarchy_pmm prc_pmm vanhanen_pmm
#> 1               NA       1            1
#> 2               NA       1            1
#> 3               NA       1            1
#> 4               NA       1            1
#> 5               NA       1            1
#> 6               NA       1            1
#> 7               NA       1            1
#> 8               NA       1            1
#> 9               NA       1            1
#> 10              NA       1            1
#> 11              NA       1            1
#> 12              NA       1            1
#> 13              NA       1            1
#> 14              NA       1            1
#> 15              NA       1            1
#> 16              NA       1            1
#> 17              NA       1            1
#> 18              NA       1            1
#> 19              NA       1            1
#> 20              NA       1            1
#> 21              NA       1            1
#> 22              NA       1            1
#> 23              NA       1            1
#> 24              NA       1            1
#> 25              NA       1            1
#> 26              NA       1            1
#> 27              NA       1            1
#> 28              NA       1            1
#> 29              NA       1            1
#> 30              NA       1            1
#> 31              NA       1            1
#> 32              NA       1            1
#> 33              NA       1            1
#> 34              NA       1            1
#> 35              NA       1            1
#> 36              NA       1            1
#> 37              NA       1            1
#>  [ reached getOption("max.print") -- omitted 9100 rows ]