Extract rater info from a UD model in a tidy format.

Extract rater info from a UD model in a tidy format.

raterinfo(model)

Arguments

model
A mirt SingleGroupClass-class model of the democracy scores.

Value

A data frame with rater information for each democracy index over the range of the latent variable theta.

Examples

replication_2011_model <- democracy_model(dplyr::matches("pmm"), verbose = FALSE)
raterinfo(replication_2011_model)

#> # A tibble: 732 x 3
#>       rater theta         info
#>       <chr> <dbl>        <dbl>
#>  1 pmm_arat  -6.0 1.215755e-06
#>  2 pmm_arat  -5.8 2.465780e-06
#>  3 pmm_arat  -5.6 5.001062e-06
#>  4 pmm_arat  -5.4 1.014308e-05
#>  5 pmm_arat  -5.2 2.057204e-05
#>  6 pmm_arat  -5.0 4.172382e-05
#>  7 pmm_arat  -4.8 8.462315e-05
#>  8 pmm_arat  -4.6 1.716292e-04
#>  9 pmm_arat  -4.4 3.480865e-04
#> 10 pmm_arat  -4.2 7.059443e-04
#> # ... with 722 more rows