The Unified Democracy Scores

The 2014, 2011, and 2010 releases of the Unified Democracy Scores by Pemstein, Meserve, and Melton (2010).

uds_2014

uds_2011

uds_2010

Format

An object of class tbl_df (inherits from tbl, data.frame) with 9850 rows and 37 columns.

Details

uds_country

The name of the country in the UDS release

uds_ccode

The correlates of war numeric code, as in the original UDS release. This is different for Yugoslavia/Serbia from the actual COW code after 2006. Use the cown variable instead for merging with COW-indexed data.

year

Calendar year.

country_name

Standardized country name.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

GWc

Gleditsch and Ward's alphabetic country code, from the Gleditsch and Ward list of independent states. For details, see Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html.

cown

The Correlates of War numeric country code. This differs from Gleditsch and Ward's numeric country code in a few cases, due to the treatment of a couple of country terminations: West Germany becomes 255 in CoW after reunification, but stays as 260 in G & W; Serbia continues as 345 in CoW after the break up of Yugoslavia, whereas it becomes 340 in GWn; Vietnam before 1948 is 816 in CoW, but 815 in G & W; and Yemen after unification is 679 in CoW, but it remains 678 on G & W, which considers it a continuation of the same state (absorbing South Yemen). Finally, Kiribati, Tonga, Tuvalu, and Nauru have different codes in G & W for reasons I cannot determine. For a full discussion of these differences, see the vignette at vignette("Spatial_and_temporal_coverage").

polity_ccode

The numeric country code of the country in the Polity dataset, which is based on, but not identical to, the CoW codes in a few cases. For a full discussion of these differences, see the vignette at vignette("Spatial_and_temporal_coverage").

year

Year. See the section on spatial and temporal organization for details on coding.

region

Region. Based on the UN region classifications. For historic states, the region is the same as the region of the successor states (or states that absorbed the territory, as for example in the case of the German principalities). In one case, Austria-Hungary, the successor states straddle two regions (Western Europe and Eastern Europe); I have opted to assign it the region "Central Europe."

continent

Continent. Based on the UN continent classifications.

GW_startdate

The entry date of the state into the international system (usually the date of independence, though there can be more than one), according to Gleditsch and Ward, or 1 January 1816, whichever is later.

GW_enddate

The exit date of the state from the international system (usually the date the state lost its independence, though there can be more than one), according to Gleditsch and Ward, or NA for countries that are still in the international system.

microstate

Whether the state is a microstate, according to Gleditsch. His tentative list of microstates is available at http://privatewww.essex.ac.uk/~ksg/statelist.html.

lat

The rough latitude of the state.

lon

The rough longitude of the state.

in_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date.

in_cow

Whether the state is in the COW system of states (that is, is independent and sovereign), for this particular date. Experimental.

mean

The posterior mean of the latent variable in the UD release.

sd

The posterior standard deviation of the latent variable in the UD release.

median

The posterior median of the latent variable in the UD release.

pct025

Unified democracy score posterior 2.5 percentile.

pct975

Unified democracy score posterior 97.5 percentile.

References

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type. Political Analysis 18 (4): 426-449.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse. http://hdl.handle.net/1902.1/PMM

See also

Other democracy: democracy, extended_uds, kailitz, original_pmm_democracy_data_1946_2008