Prepare democracy data before replicating the UDS model

This function is designed to take the democracy data included in this package and put it in a form suitable for use with the mirt package to replicate the UDS model. It takes a data frame and tries to determine, from the colum names, which variables contain democracy scores.

prepare_data(data)

Arguments

data

A dataset of democracy scores. For the function to do anything, the column names must contain at least one of the following strings: arat, blm, bmr, bollen, doorenspleet, eiu, e_v2x, gwf, hadenius, kailitz, lied, munck, pacl, peps, poliarchy, polity, prc, przeworski, svolik, ulfelder, utip, v2x, vanhanen_democratization, vanhanen_pmm, or wahman_teorell_hadenius. For details of these variables, see the documentation for democracy.

Value

A data frame with the transformed scores, if any.

Details

If the column names contain the strings arat, blm, bmr, bollen, doorenspleet, eiu, e_v2x, gwf, hadenius, kailitz, lied, munck, pacl, peps, poliarchy, polity, prc, svolik, ulfelder, v2x, vanhanen_pmm, vanhanen_democratization, or wahman_teorell_hadenius, the function performs the following transformations by default:

arat: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function cuts Arat (1991)'s 0-109 democracy score into 7 intervals with the following cutoffs: 50, 60, 70, 80, 90, and 100. The resulting score is ordinal from 1 to 8.

bollen: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function cuts Bollen's (2001)'s 0-100 democracy score into 10 intervals with the following cutoffs: 10,20,30,40,50,60,70,80, and 90. The resulting score is ordinal from 1 to 10.

eiu: If the Economist Intelligence Unit's index of democracy is included in the file, the function rounds the 0-1 value to the first decimal place and cuts the result into 10 categories. The resulting score is ordinal from 1 to 11.

hadenius: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function cuts Hadenius (1992)'s 0-10 democracy score into 8 intervals with the following cutoffs: 1, 2,3,4, 7, 8, and 9. The resulting score is ordinal from 1 to 8.

munck: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function cuts Munck's (2009)'s 0-1 democracy score into 4 intervals with the following cutoffs: 0.5,0.5,0.75, and 0.99. The resulting score is ordinal from 1 to 4.

peps: If any of the variants of the Participation-Enhanced Polity Score (Moon et al 2006) is included in the file, the function rounds its value (eliminates the decimal) and then transforms it into an ordinal measure from 1 to 21.

polity: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function takes the polity scores and puts NA for any values below -10, and then transforms it into an ordinal measure from 1 to 21.

v2x: If any of the v2x_ continuous indexes of democracy from the V-Dem dataset (Coppedge et al 2015) are included in the file, the function cuts them into 20 categories. The resulting score is ordinal from 1 to 20.

vanhanen_democratization or vanhanen_pmm: Following Pemstein, Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013), the function cuts Vanhanen's (2012)'s index of democratization into 8 intervals with the following cutoffs: 5,10,15,20,25,30, and 35. The resulting score is ordinal from 1 to 8.

The function also recognizes the following column names (or partial column names - it also recognizes, e.g., blm_pmm) as measures of democracy: blm (from Bowman, Lehoucq, and Mahoney 2005), bmr (from Boix, Miller, and Rosato 2012), doorenspleet (from Doorenspleet 2000), e_v2x (the "ordinal" indexes from the V-dem project, Coppedge at al 2015), freedomhouse (from Freedom House - freedom scale must be reversed so that "more freedom" is higher), gwf (from Geddes, Wright, and Frantz 2014 - the dichotomous democracy indicator only), kailitz (from Kailitz 2013 - democracy/non-democracy indicator), lied (from Skaaning, Gerring, and Bartusevicius 2015), mainwaring (from Mainwaring and Perez Linan 2008), magaloni (from Magaloni, Min, Chu 2013 - democracy/non-democracy indicator), pacl (from Cheibub, Gandhi, and Vreeland 2010), pitf (from Goldstone et al 2010 or Taylor and Ulfelder 2015), poliarchy (from Coppedge and Reinicke 1991), prc (from Gasiorowski 1996 or Reich 2002), przeworski (from Przeworski 2010), svolik (from Svolik 2012, democracy/dictatorship indicator only), ulfelder (from Ulfelder 2012), utip (from Hsu 2008), and wahman_teorell_hadenius (from Wahman, Teorell, and Hadenius 2013). In each of these cases the function transforms the values of these scores by running as.numeric(unclass(factor(x))), which transforms them into ordinal variables from 1 to the number of categories.

For details of these scores, see the documentation for democracy.

It is also possible to change these defaults.

Note

Warning! The function does not perform any sanity checks. It will try to transform anything that has the right name. You should always check the results make sense.

References

Arat, Zehra F. 1991. Democracy and human rights in developing countries. Boulder: Lynne Rienner Publishers.

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete Data Set of Political Regimes, 1800-2007. Comparative Political Studies 46 (12): 1523-1554. Original data available at https://sites.google.com/site/mkmtwo/democracy-v2.0.dta?attredirects=0

Bollen, Kenneth A. 2001. "Cross-National Indicators of Liberal Democracy, 1950-1990." 2nd ICPSR version. Chapel Hill, NC: University of North Carolina, 1998. Ann Arbor, MI: Inter-university Consortium for Political and Social Research, 2001. Original data available at http://webapp.icpsr.umich.edu/cocoon/ICPSR-STUDY/02532.xml.

Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America. Comparative Political Studies 38 (8): 939-970. http://cps.sagepub.com/content/38/8/939. Data available at http://www.blmdemocracy.gatech.edu/.

Cheibub, Jose Antonio, Jennifer Gandhi, and James Raymond Vreeland. 2010. "Democracy and Dictatorship Revisited." Public Choice. 143(1):67-101. Original data available at https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited.

Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning, and Jan Teorell, with David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Kelly McMann, Pamela Paxton, Daniel Pemstein, Jeffrey Staton, Brigitte Zimmerman, Frida Andersson, Valeriya Mechkova, Farhad Miri. 2015. V-Dem Codebook v5. Varieties of Democracy (V-Dem) Project. Original data available at https://v-dem.net/en/data/.

Coppedge, Michael and Wolfgang H. Reinicke. 1991. Measuring Polyarchy. In On Measuring Democracy: Its Consequences and Concomitants, ed. Alex Inkeles. New Brunswuck, NJ: Transaction pp. 47-68.

Doorenspleet, Renske. 2000. Reassessing the Three Waves of Democratization. World Politics 52 (03): 384-406. DOI: 10.1017/S0043887100016580. http://dx.doi.org/10.1017/S0043887100016580.

Economist Intelligence Unit. 2012. Democracy Index 2012: Democracy at a Standstill.

Freedom House. 2015. "Freedom in the World." Original data available at http://www.freedomhouse.org.

Gasiorowski, Mark J. 1996. "An Overview of the Political Regime Change Dataset." Comparative Political Studies 29(4):469-483.

Geddes, Barbara, Joseph Wright, and Erica Frantz. 2014. Autocratic Breakdown and Regime Transitions: A New Data Set. Perspectives on Politics 12 (1): 313-331. Original data available at http://dictators.la.psu.edu/.

Goldstone, Jack, Robert Bates, David Epstein, Ted Gurr, Michael Lustik, Monty Marshall, Jay Ulfelder, and Mark Woodward. 2010. A Global Model for Forecasting Political Instability. American Journal of Political Science 54 (1): 190-208. DOI:10.1111/j.1540-5907.2009.00426.x

Hadenius, Axel. 1992. Democracy and Development. Cambridge: Cambridge University Press.

Hadenius, Axel & Jan Teorell. 2007. "Pathways from Authoritarianism", Journal of Democracy 18(1): 143-156.

Hsu, Sara "The Effect of Political Regimes on Inequality, 1963-2002," UTIP Working Paper No. 53 (2008), http://utip.gov.utexas.edu/papers/utip_53.pdf. Data available for download at http://utip.gov.utexas.edu/data/.

Kailitz, Steffen. 2013. Classifying political regimes revisited: legitimation and durability. Democratization 20 (1): 39-60. Original data available at http://dx.doi.org/10.1080/13510347.2013.738861.

Mainwaring, Scott, Daniel Brinks, and Anibal Perez Linan. 2008. "Political Regimes in Latin America, 1900-2007." Original data available from http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf.

Magaloni, Beatriz, Jonathan Chu, and Eric Min. 2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford University. Original data and codebook available at http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/

Marshall, Monty G., Ted Robert Gurr, and Keith Jaggers. 2012. "Polity IV: Political Regime Characteristics and Transitions, 1800-2012." Updated to 2015. Original data available from http://www.systemicpeace.org/polity/polity4.htm.

Moon, Bruce E., Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of Democracy" Studies in Comparative International Development 42, 2 (Summer, 2006). The complete dataset is available here: http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm.

Munck, Gerardo L. 2009. Measuring Democracy: A Bridge Between Scholarship and Politics. Baltimore: Johns Hopkins University Press.

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type. Political Analysis 18 (4): 426-449.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse. http://hdl.handle.net/1902.1/PMM

Przeworski, Adam et al. 2013. Political Institutions and Political Events (PIPE) Data Set. Department of Politics, New York University. https://sites.google.com/a/nyu.edu/adam-przeworski/home/data

Reich, G. 2002. Categorizing Political Regimes: New Data for Old Problems. Democratization 9 (4): 1-24. http://www.tandfonline.com/doi/pdf/10.1080/714000289.

Skaaning, Svend-Erik, John Gerring, and Henrikas Bartusevicius. 2015. A Lexical Index of Electoral Democracy. Comparative Political Studies 48 (12): 1491-1525. Original data available from http://thedata.harvard.edu/dvn/dv/skaaning.

Svolik, Milan. 2012. The Politics of Authoritarian Rule. Cambridge and New York: Cambridge University Press. Original data available from http://campuspress.yale.edu/svolik/the-politics-of-authoritarian-rule/.

Taylor, Sean J. and Ulfelder, Jay, A Measurement Error Model of Dichotomous Democracy Status (May 20, 2015). Available at SSRN: http://ssrn.com/abstract=2726962 or http://dx.doi.org/10.2139/ssrn.2726962

Ulfelder, Jay. 2012. "Democracy/Autocracy Data Set." In: Harvard Dataverse. http://hdl.handle.net/1902.1/18836.

Vanhanen, Tatu. 2012. "FSD1289 Measures of Democracy 1810-2012." Original data available from http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime types revisited: updated data in comparative perspective. Contemporary Politics 19 (1): 19-34.

Examples

summary(democracy)

#>  country_name            GWn            GWc                 cown      
#>  Length:24175       Min.   :  2.0   Length:24175       Min.   :  2.0  
#>  Class :character   1st Qu.:221.0   Class :character   1st Qu.:221.0  
#>  Mode  :character   Median :432.0   Mode  :character   Median :432.0  
#>                     Mean   :440.5                      Mean   :440.4  
#>                     3rd Qu.:651.0                      3rd Qu.:651.0  
#>                     Max.   :990.0                      Max.   :990.0  
#>                                                                       
#>   polity_ccode        year       GW_startdate          GW_enddate        
#>  Min.   :  2.0   Min.   :1741   Min.   :1816-01-01   Min.   :1830-07-05  
#>  1st Qu.:221.0   1st Qu.:1900   1st Qu.:1816-01-01   1st Qu.:1871-01-17  
#>  Median :432.0   Median :1945   Median :1907-09-01   Median :1893-01-01  
#>  Mean   :440.4   Mean   :1935   Mean   :1894-08-19   Mean   :1902-01-19  
#>  3rd Qu.:651.0   3rd Qu.:1984   3rd Qu.:1960-10-01   3rd Qu.:1915-10-01  
#>  Max.   :990.0   Max.   :2015   Max.   :2011-07-09   Max.   :2006-06-04  
#>                                                      NA's   :21097       
#>     region           continent         microstate           lat        
#>  Length:24175       Length:24175       Mode :logical   Min.   :-40.90  
#>  Class :character   Class :character   FALSE:23058     1st Qu.:  4.21  
#>  Mode  :character   Mode  :character   TRUE :1117      Median : 19.86  
#>                                        NA's :0         Mean   : 19.27  
#>                                                        3rd Qu.: 39.40  
#>                                                        Max.   : 64.96  
#>                                                                        
#>       lon          in_system         in_cow           arat_pmm    
#>  Min.   :-175.20   Mode :logical   Mode :logical   Min.   : 29.0  
#>  1st Qu.: -11.78   FALSE:5922      FALSE:7261      1st Qu.: 58.0  
#>  Median :  16.37   TRUE :18253     TRUE :16914     Median : 69.0  
#>  Mean   :  14.89   NA's :0         NA's :0         Mean   : 73.2  
#>  3rd Qu.:  46.87                                   3rd Qu.: 92.0  
#>  Max.   : 178.68                                   Max.   :109.0  
#>                                                    NA's   :20302  
#>       blm           blm_pmm      bmr_democracy   bmr_democracy_omitteddata
#>  Min.   :0.000   Min.   :0.00    Min.   :0.000   Min.   :0.000            
#>  1st Qu.:0.000   1st Qu.:0.00    1st Qu.:0.000   1st Qu.:0.000            
#>  Median :0.000   Median :0.00    Median :0.000   Median :0.000            
#>  Mean   :0.252   Mean   :0.36    Mean   :0.319   Mean   :0.321            
#>  3rd Qu.:0.500   3rd Qu.:0.50    3rd Qu.:1.000   3rd Qu.:1.000            
#>  Max.   :1.000   Max.   :1.00    Max.   :1.000   Max.   :1.000            
#>  NA's   :23670   NA's   :23900   NA's   :7189    NA's   :7414             
#>       bnr          bollen_pmm      doorenspleet        eiu       
#>  Min.   :0.000   Min.   :  0.00   Min.   :1.000   Min.   :0.000  
#>  1st Qu.:0.000   1st Qu.: 22.84   1st Qu.:1.000   1st Qu.:0.279  
#>  Median :0.000   Median : 53.59   Median :1.000   Median :0.448  
#>  Mean   :0.347   Mean   : 55.46   Mean   :1.179   Mean   :0.467  
#>  3rd Qu.:1.000   3rd Qu.: 90.95   3rd Qu.:1.000   3rd Qu.:0.649  
#>  Max.   :1.000   Max.   :100.00   Max.   :2.000   Max.   :0.965  
#>  NA's   :13823   NA's   :23665    NA's   :11166   NA's   :21774  
#>     exconst            exrec          freedomhouse   freedomhouse_electoral
#>  Min.   :-88.000   Min.   :-88.000   Min.   :1.000   Min.   :0.000         
#>  1st Qu.:  1.000   1st Qu.:  2.000   1st Qu.:2.500   1st Qu.:0.000         
#>  Median :  3.000   Median :  3.000   Median :4.500   Median :1.000         
#>  Mean   :  0.088   Mean   :  0.862   Mean   :4.259   Mean   :0.599         
#>  3rd Qu.:  7.000   3rd Qu.:  8.000   3rd Qu.:6.500   3rd Qu.:1.000         
#>  Max.   :  7.000   Max.   :  8.000   Max.   :7.000   Max.   :1.000         
#>  NA's   :7122      NA's   :7323      NA's   :16496   NA's   :19103         
#>  freedomhouse_pmm      gwf         hadenius_pmm    kailitz_binary 
#>  Min.   :1.00     Min.   :1.000   Min.   : 0.000   Min.   :0.000  
#>  1st Qu.:2.50     1st Qu.:1.000   1st Qu.: 1.500   1st Qu.:0.000  
#>  Median :4.00     Median :1.000   Median : 3.100   Median :0.000  
#>  Mean   :4.15     Mean   :1.429   Mean   : 4.509   Mean   :0.412  
#>  3rd Qu.:6.00     3rd Qu.:2.000   3rd Qu.: 8.300   3rd Qu.:1.000  
#>  Max.   :7.00     Max.   :2.000   Max.   :10.000   Max.   :1.000  
#>  NA's   :17737    NA's   :14780   NA's   :24046    NA's   :14569  
#>   kailitz_tri         lied       lied_accountable lied_electoral 
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000    Min.   :0.000  
#>  1st Qu.:0.000   1st Qu.:1.000   1st Qu.:0.000    1st Qu.:1.000  
#>  Median :1.000   Median :3.000   Median :1.000    Median :2.000  
#>  Mean   :0.989   Mean   :2.783   Mean   :0.944    Mean   :1.382  
#>  3rd Qu.:2.000   3rd Qu.:6.000   3rd Qu.:2.000    3rd Qu.:2.000  
#>  Max.   :2.000   Max.   :6.000   Max.   :2.000    Max.   :2.000  
#>  NA's   :14569   NA's   :5646    NA's   :5646     NA's   :5646   
#>  lied_inclusive magaloni_democ_binary magaloni_regime_tri   mainwaring    
#>  Min.   :0.00   Min.   :0.000         Min.   :1.000       Min.   :-1.000  
#>  1st Qu.:0.00   1st Qu.:0.000         1st Qu.:1.000       1st Qu.:-1.000  
#>  Median :2.00   Median :0.000         Median :2.000       Median :-1.000  
#>  Mean   :1.17   Mean   :0.424         Mean   :2.011       Mean   :-0.325  
#>  3rd Qu.:2.00   3rd Qu.:1.000         3rd Qu.:3.000       3rd Qu.: 0.000  
#>  Max.   :2.00   Max.   :1.000         Max.   :3.000       Max.   : 1.000  
#>  NA's   :5646   NA's   :13887         NA's   :13887       NA's   :22019   
#>  mainwaring_pmm     munck_pmm          pacl          pacl_pmm    
#>  Min.   :-1.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:-1.000   1st Qu.:0.750   1st Qu.:0.000   1st Qu.:0.000  
#>  Median : 0.000   Median :1.000   Median :0.000   Median :0.000  
#>  Mean   : 0.122   Mean   :0.838   Mean   :0.438   Mean   :0.436  
#>  3rd Qu.: 1.000   3rd Qu.:1.000   3rd Qu.:1.000   3rd Qu.:1.000  
#>  Max.   : 1.000   Max.   :1.000   Max.   :1.000   Max.   :1.000  
#>  NA's   :23340    NA's   :23833   NA's   :15060   NA's   :15108  
#>      PEPS1i            PEPS1q            PEPS1v            PEPS2i      
#>  Min.   :-10.000   Min.   :-10.000   Min.   :-10.000   Min.   :-9.590  
#>  1st Qu.: -9.000   1st Qu.: -9.000   1st Qu.: -7.000   1st Qu.: 0.170  
#>  Median : -7.000   Median : -6.240   Median : -4.000   Median : 5.000  
#>  Mean   : -3.199   Mean   : -3.176   Mean   : -2.869   Mean   : 3.716  
#>  3rd Qu.:  3.700   3rd Qu.:  3.480   3rd Qu.:  1.500   3rd Qu.: 7.450  
#>  Max.   : 10.000   Max.   : 10.000   Max.   : 10.000   Max.   :10.000  
#>  NA's   :15012     NA's   :14736     NA's   :10086     NA's   :20050   
#>      PEPS2q            PEPS2v             pitf        pitf_binary   
#>  Min.   :-10.000   Min.   :-10.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.: -8.500   1st Qu.: -7.500   1st Qu.:1.000   1st Qu.:1.000  
#>  Median : -1.455   Median : -3.550   Median :2.000   Median :1.000  
#>  Mean   : -1.126   Mean   : -2.427   Mean   :2.425   Mean   :1.363  
#>  3rd Qu.:  5.890   3rd Qu.:  1.937   3rd Qu.:4.000   3rd Qu.:2.000  
#>  Max.   : 10.000   Max.   : 10.000   Max.   :5.000   Max.   :2.000  
#>  NA's   :17285     NA's   :10934     NA's   :8089    NA's   :8089   
#>     polcomp            polity         polity_pmm         polity2       
#>  Min.   :-88.000   Min.   :-88.00   Min.   :-10.000   Min.   :-10.000  
#>  1st Qu.:  1.000   1st Qu.: -7.00   1st Qu.: -7.000   1st Qu.: -7.000  
#>  Median :  6.000   Median : -3.00   Median : -1.000   Median : -3.000  
#>  Mean   :  1.302   Mean   : -4.08   Mean   :  0.129   Mean   : -0.596  
#>  3rd Qu.:  7.000   3rd Qu.:  7.00   3rd Qu.:  8.000   3rd Qu.:  7.000  
#>  Max.   : 10.000   Max.   : 10.00   Max.   : 10.000   Max.   : 10.000  
#>  NA's   :7248      NA's   :7122     NA's   :16125     NA's   :7357     
#>     Polity3        polyarchy_contestation polyarchy_pmm    polyarchy_reversed
#>  Min.   :-10.000   Min.   :1.000          Min.   : 0.000   Min.   : 0.000    
#>  1st Qu.: -7.000   1st Qu.:3.000          1st Qu.: 3.000   1st Qu.: 3.000    
#>  Median : -3.000   Median :6.000          Median : 7.000   Median : 7.000    
#>  Mean   : -1.127   Mean   :5.821          Mean   : 6.329   Mean   : 6.353    
#>  3rd Qu.:  6.000   3rd Qu.:9.000          3rd Qu.:10.000   3rd Qu.:10.000    
#>  Max.   : 10.000   Max.   :9.000          Max.   :10.000   Max.   :10.000    
#>  NA's   :9367      NA's   :23818          NA's   :23822    NA's   :23818     
#>       prc         prc_notrans       prc_pmm        przeworski   
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :1.000   Median :1.000   Median :1.000   Median :2.000  
#>  Mean   :2.002   Mean   :2.002   Mean   :2.147   Mean   :1.788  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:4.000   3rd Qu.:2.000  
#>  Max.   :4.000   Max.   :4.000   Max.   :4.000   Max.   :3.000  
#>  NA's   :13144   NA's   :13292   NA's   :18173   NA's   :10122  
#>      svolik         ulfelder     utip_dichotomous utip_dichotomous_strict
#>  Min.   :1.000   Min.   :0.000   Min.   :0.000    Min.   :0.000          
#>  1st Qu.:1.000   1st Qu.:0.000   1st Qu.:0.000    1st Qu.:0.000          
#>  Median :1.000   Median :0.000   Median :1.000    Median :0.000          
#>  Mean   :1.439   Mean   :0.412   Mean   :0.515    Mean   :0.477          
#>  3rd Qu.:2.000   3rd Qu.:1.000   3rd Qu.:1.000    3rd Qu.:1.000          
#>  Max.   :2.000   Max.   :1.000   Max.   :1.000    Max.   :1.000          
#>  NA's   :15403   NA's   :16491   NA's   :18132    NA's   :18132          
#>  utip_trichotomous    v2x_api       v2x_delibdem    v2x_egaldem   
#>  Min.   :0.000     Min.   :0.017   Min.   :0.000   Min.   :0.007  
#>  1st Qu.:0.000     1st Qu.:0.194   1st Qu.:0.014   1st Qu.:0.062  
#>  Median :1.000     Median :0.414   Median :0.071   Median :0.147  
#>  Mean   :0.992     Mean   :0.469   Mean   :0.213   Mean   :0.246  
#>  3rd Qu.:2.000     3rd Qu.:0.774   3rd Qu.:0.358   3rd Qu.:0.347  
#>  Max.   :2.000     Max.   :0.983   Max.   :0.929   Max.   :0.925  
#>  NA's   :18132     NA's   :8103    NA's   :8211    NA's   :8103   
#>    v2x_libdem       v2x_mpi      v2x_partipdem   v2x_polyarchy  
#>  Min.   :0.011   Min.   :0.000   Min.   :0.000   Min.   :0.008  
#>  1st Qu.:0.077   1st Qu.:0.000   1st Qu.:0.048   1st Qu.:0.097  
#>  Median :0.149   Median :0.001   Median :0.106   Median :0.209  
#>  Mean   :0.261   Mean   :0.176   Mean   :0.200   Mean   :0.323  
#>  3rd Qu.:0.385   3rd Qu.:0.292   3rd Qu.:0.306   3rd Qu.:0.535  
#>  Max.   :0.928   Max.   :0.934   Max.   :0.840   Max.   :0.958  
#>  NA's   :8103    NA's   :8103    NA's   :8112    NA's   :8103   
#>  vanhanen_competition vanhanen_democratization vanhanen_participation
#>  Min.   : 0.00        Min.   : 0.000           Min.   : 0.0          
#>  1st Qu.: 0.00        1st Qu.: 0.000           1st Qu.: 0.0          
#>  Median :20.00        Median : 1.100           Median :14.0          
#>  Mean   :25.22        Mean   : 8.434           Mean   :21.1          
#>  3rd Qu.:49.10        3rd Qu.:15.400           3rd Qu.:40.4          
#>  Max.   :70.00        Max.   :49.000           Max.   :71.0          
#>  NA's   :8440         NA's   :8440             NA's   :8440          
#>   vanhanen_pmm   wahman_teorell_hadenius
#>  Min.   : 0.00   Min.   :0.000          
#>  1st Qu.: 0.00   1st Qu.:0.000          
#>  Median : 5.90   Median :0.000          
#>  Mean   :11.31   Mean   :0.424          
#>  3rd Qu.:20.70   3rd Qu.:1.000          
#>  Max.   :49.00   Max.   :1.000          
#>  NA's   :15210   NA's   :17482          

summary(prepare_data(democracy))

#>  country_name            GWn            GWc                 cown      
#>  Length:24175       Min.   :  2.0   Length:24175       Min.   :  2.0  
#>  Class :character   1st Qu.:221.0   Class :character   1st Qu.:221.0  
#>  Mode  :character   Median :432.0   Mode  :character   Median :432.0  
#>                     Mean   :440.5                      Mean   :440.4  
#>                     3rd Qu.:651.0                      3rd Qu.:651.0  
#>                     Max.   :990.0                      Max.   :990.0  
#>                                                                       
#>   polity_ccode        year       GW_startdate          GW_enddate        
#>  Min.   :  1.0   Min.   :1741   Min.   :1816-01-01   Min.   :1830-07-05  
#>  1st Qu.: 44.0   1st Qu.:1900   1st Qu.:1816-01-01   1st Qu.:1871-01-17  
#>  Median :112.0   Median :1945   Median :1907-09-01   Median :1893-01-01  
#>  Mean   :111.3   Mean   :1935   Mean   :1894-08-19   Mean   :1902-01-19  
#>  3rd Qu.:169.0   3rd Qu.:1984   3rd Qu.:1960-10-01   3rd Qu.:1915-10-01  
#>  Max.   :232.0   Max.   :2015   Max.   :2011-07-09   Max.   :2006-06-04  
#>                                                      NA's   :21097       
#>     region           continent         microstate           lat        
#>  Length:24175       Length:24175       Mode :logical   Min.   :-40.90  
#>  Class :character   Class :character   FALSE:23058     1st Qu.:  4.21  
#>  Mode  :character   Mode  :character   TRUE :1117      Median : 19.86  
#>                                        NA's :0         Mean   : 19.27  
#>                                                        3rd Qu.: 39.40  
#>                                                        Max.   : 64.96  
#>                                                                        
#>       lon          in_system         in_cow           arat_pmm    
#>  Min.   :-175.20   Mode :logical   Mode :logical   Min.   :1.000  
#>  1st Qu.: -11.78   FALSE:5922      FALSE:7261      1st Qu.:2.000  
#>  Median :  16.37   TRUE :18253     TRUE :16914     Median :3.000  
#>  Mean   :  14.89   NA's :0         NA's :0         Mean   :3.878  
#>  3rd Qu.:  46.87                                   3rd Qu.:6.000  
#>  Max.   : 178.68                                   Max.   :7.000  
#>                                                    NA's   :20302  
#>       blm           blm_pmm      bmr_democracy   bmr_democracy_omitteddata
#>  Min.   :1.000   Min.   :1.00    Min.   :1.000   Min.   :1.000            
#>  1st Qu.:1.000   1st Qu.:1.00    1st Qu.:1.000   1st Qu.:1.000            
#>  Median :1.000   Median :1.00    Median :1.000   Median :1.000            
#>  Mean   :1.505   Mean   :1.72    Mean   :1.319   Mean   :1.321            
#>  3rd Qu.:2.000   3rd Qu.:2.00    3rd Qu.:2.000   3rd Qu.:2.000            
#>  Max.   :3.000   Max.   :3.00    Max.   :2.000   Max.   :2.000            
#>  NA's   :23670   NA's   :23900   NA's   :7189    NA's   :7414             
#>       bnr          bollen_pmm      doorenspleet        eiu        
#>  Min.   :0.000   Min.   : 1.000   Min.   :1.000   Min.   : 1.000  
#>  1st Qu.:0.000   1st Qu.: 3.000   1st Qu.:1.000   1st Qu.: 4.000  
#>  Median :0.000   Median : 6.000   Median :1.000   Median : 5.000  
#>  Mean   :0.347   Mean   : 6.006   Mean   :1.179   Mean   : 5.665  
#>  3rd Qu.:1.000   3rd Qu.:10.000   3rd Qu.:1.000   3rd Qu.: 7.000  
#>  Max.   :1.000   Max.   :10.000   Max.   :2.000   Max.   :11.000  
#>  NA's   :13823   NA's   :23665    NA's   :11166   NA's   :21774   
#>     exconst            exrec          freedomhouse    freedomhouse_electoral
#>  Min.   :-88.000   Min.   :-88.000   Min.   : 1.000   Min.   :1.000         
#>  1st Qu.:  1.000   1st Qu.:  2.000   1st Qu.: 4.000   1st Qu.:1.000         
#>  Median :  3.000   Median :  3.000   Median : 8.000   Median :2.000         
#>  Mean   :  0.088   Mean   :  0.862   Mean   : 7.518   Mean   :1.599         
#>  3rd Qu.:  7.000   3rd Qu.:  8.000   3rd Qu.:12.000   3rd Qu.:2.000         
#>  Max.   :  7.000   Max.   :  8.000   Max.   :13.000   Max.   :2.000         
#>  NA's   :7122      NA's   :7323      NA's   :16496    NA's   :19103         
#>  freedomhouse_pmm      gwf         hadenius_pmm   kailitz_binary 
#>  Min.   : 1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.: 4.000   1st Qu.:1.000   1st Qu.:2.000   1st Qu.:1.000  
#>  Median : 7.000   Median :1.000   Median :4.000   Median :1.000  
#>  Mean   : 7.301   Mean   :1.429   Mean   :4.279   Mean   :1.412  
#>  3rd Qu.:11.000   3rd Qu.:2.000   3rd Qu.:7.000   3rd Qu.:2.000  
#>  Max.   :13.000   Max.   :2.000   Max.   :8.000   Max.   :2.000  
#>  NA's   :17737    NA's   :14780   NA's   :24046   NA's   :14569  
#>   kailitz_tri         lied       lied_accountable lied_electoral 
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000    Min.   :1.000  
#>  1st Qu.:1.000   1st Qu.:2.000   1st Qu.:1.000    1st Qu.:2.000  
#>  Median :2.000   Median :4.000   Median :2.000    Median :3.000  
#>  Mean   :1.989   Mean   :3.783   Mean   :1.944    Mean   :2.382  
#>  3rd Qu.:3.000   3rd Qu.:7.000   3rd Qu.:3.000    3rd Qu.:3.000  
#>  Max.   :3.000   Max.   :7.000   Max.   :3.000    Max.   :3.000  
#>  NA's   :14569   NA's   :5646    NA's   :5646     NA's   :5646   
#>  lied_inclusive magaloni_democ_binary magaloni_regime_tri   mainwaring   
#>  Min.   :1.00   Min.   :1.000         Min.   :1.000       Min.   :1.000  
#>  1st Qu.:1.00   1st Qu.:1.000         1st Qu.:1.000       1st Qu.:1.000  
#>  Median :3.00   Median :1.000         Median :2.000       Median :1.000  
#>  Mean   :2.17   Mean   :1.424         Mean   :2.011       Mean   :1.675  
#>  3rd Qu.:3.00   3rd Qu.:2.000         3rd Qu.:3.000       3rd Qu.:2.000  
#>  Max.   :3.00   Max.   :2.000         Max.   :3.000       Max.   :3.000  
#>  NA's   :5646   NA's   :13887         NA's   :13887       NA's   :22019  
#>  mainwaring_pmm    munck_pmm          pacl          pacl_pmm    
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:1.000   1st Qu.:3.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :2.000   Median :4.000   Median :1.000   Median :1.000  
#>  Mean   :2.122   Mean   :3.333   Mean   :1.438   Mean   :1.436  
#>  3rd Qu.:3.000   3rd Qu.:4.000   3rd Qu.:2.000   3rd Qu.:2.000  
#>  Max.   :3.000   Max.   :4.000   Max.   :2.000   Max.   :2.000  
#>  NA's   :23340   NA's   :23833   NA's   :15060   NA's   :15108  
#>      PEPS1i           PEPS1q           PEPS1v           PEPS2i     
#>  Min.   : 1.000   Min.   : 1.000   Min.   : 1.000   Min.   : 1.00  
#>  1st Qu.: 2.000   1st Qu.: 2.000   1st Qu.: 4.000   1st Qu.:11.00  
#>  Median : 4.000   Median : 5.000   Median : 7.000   Median :16.00  
#>  Mean   : 7.804   Mean   : 7.828   Mean   : 8.117   Mean   :14.71  
#>  3rd Qu.:15.000   3rd Qu.:14.000   3rd Qu.:12.000   3rd Qu.:18.00  
#>  Max.   :21.000   Max.   :21.000   Max.   :21.000   Max.   :21.00  
#>  NA's   :15012    NA's   :14736    NA's   :10086    NA's   :20050  
#>      PEPS2q           PEPS2v           pitf        pitf_binary   
#>  Min.   : 1.000   Min.   : 1.0    Min.   :1.000   Min.   :1.000  
#>  1st Qu.: 3.000   1st Qu.: 3.0    1st Qu.:1.000   1st Qu.:1.000  
#>  Median :10.000   Median : 7.0    Median :2.000   Median :1.000  
#>  Mean   : 9.917   Mean   : 8.6    Mean   :2.425   Mean   :1.363  
#>  3rd Qu.:17.000   3rd Qu.:13.0    3rd Qu.:4.000   3rd Qu.:2.000  
#>  Max.   :21.000   Max.   :21.0    Max.   :5.000   Max.   :2.000  
#>  NA's   :17285    NA's   :10934   NA's   :8089    NA's   :8089   
#>     polcomp            polity        polity_pmm       polity2    
#>  Min.   :-88.000   Min.   : 1.00   Min.   : 1.00   Min.   : 1.0  
#>  1st Qu.:  1.000   1st Qu.: 4.00   1st Qu.: 4.00   1st Qu.: 4.0  
#>  Median :  6.000   Median : 8.00   Median :10.00   Median : 8.0  
#>  Mean   :  1.302   Mean   :10.41   Mean   :11.13   Mean   :10.4  
#>  3rd Qu.:  7.000   3rd Qu.:18.00   3rd Qu.:19.00   3rd Qu.:18.0  
#>  Max.   : 10.000   Max.   :21.00   Max.   :21.00   Max.   :21.0  
#>  NA's   :7248      NA's   :7888    NA's   :16125   NA's   :7357  
#>     Polity3       polyarchy_contestation polyarchy_pmm    polyarchy_reversed
#>  Min.   : 1.000   Min.   :1.000          Min.   : 1.000   Min.   : 1.000    
#>  1st Qu.: 4.000   1st Qu.:3.000          1st Qu.: 4.000   1st Qu.: 4.000    
#>  Median : 8.000   Median :6.000          Median : 8.000   Median : 8.000    
#>  Mean   : 9.873   Mean   :5.821          Mean   : 7.329   Mean   : 7.353    
#>  3rd Qu.:17.000   3rd Qu.:9.000          3rd Qu.:11.000   3rd Qu.:11.000    
#>  Max.   :21.000   Max.   :9.000          Max.   :11.000   Max.   :11.000    
#>  NA's   :9367     NA's   :23818          NA's   :23822    NA's   :23818     
#>       prc         prc_notrans       prc_pmm        przeworski   
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:2.000  
#>  Median :1.000   Median :1.000   Median :1.000   Median :3.000  
#>  Mean   :2.002   Mean   :1.621   Mean   :2.147   Mean   :2.788  
#>  3rd Qu.:3.000   3rd Qu.:2.000   3rd Qu.:4.000   3rd Qu.:3.000  
#>  Max.   :4.000   Max.   :3.000   Max.   :4.000   Max.   :4.000  
#>  NA's   :13144   NA's   :13292   NA's   :18173   NA's   :10122  
#>      svolik         ulfelder     utip_dichotomous utip_dichotomous_strict
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000    Min.   :1.000          
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000    1st Qu.:1.000          
#>  Median :1.000   Median :1.000   Median :2.000    Median :1.000          
#>  Mean   :1.439   Mean   :1.412   Mean   :1.515    Mean   :1.477          
#>  3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:2.000    3rd Qu.:2.000          
#>  Max.   :2.000   Max.   :2.000   Max.   :2.000    Max.   :2.000          
#>  NA's   :15403   NA's   :16491   NA's   :18132    NA's   :18132          
#>  utip_trichotomous    v2x_api        v2x_delibdem     v2x_egaldem    
#>  Min.   :1.000     Min.   : 1.000   Min.   : 1.000   Min.   : 1.000  
#>  1st Qu.:1.000     1st Qu.: 4.000   1st Qu.: 1.000   1st Qu.: 2.000  
#>  Median :2.000     Median : 9.000   Median : 2.000   Median : 4.000  
#>  Mean   :1.992     Mean   : 9.859   Mean   : 5.173   Mean   : 5.718  
#>  3rd Qu.:3.000     3rd Qu.:16.000   3rd Qu.: 8.000   3rd Qu.: 8.000  
#>  Max.   :3.000     Max.   :20.000   Max.   :20.000   Max.   :20.000  
#>  NA's   :18132     NA's   :8103     NA's   :8211     NA's   :8103    
#>    v2x_libdem        v2x_mpi       v2x_partipdem    v2x_polyarchy   
#>  Min.   : 1.000   Min.   : 1.000   Min.   : 1.000   Min.   : 1.000  
#>  1st Qu.: 2.000   1st Qu.: 1.000   1st Qu.: 2.000   1st Qu.: 2.000  
#>  Median : 4.000   Median : 1.000   Median : 3.000   Median : 5.000  
#>  Mean   : 5.964   Mean   : 4.557   Mean   : 5.268   Mean   : 7.137  
#>  3rd Qu.: 9.000   3rd Qu.: 7.000   3rd Qu.: 8.000   3rd Qu.:12.000  
#>  Max.   :20.000   Max.   :20.000   Max.   :20.000   Max.   :20.000  
#>  NA's   :8103     NA's   :8103     NA's   :8112     NA's   :8103    
#>  vanhanen_competition vanhanen_democratization vanhanen_participation
#>  Min.   : 0.00        Min.   :1.000            Min.   : 0.0          
#>  1st Qu.: 0.00        1st Qu.:1.000            1st Qu.: 0.0          
#>  Median :20.00        Median :1.000            Median :14.0          
#>  Mean   :25.22        Mean   :2.422            Mean   :21.1          
#>  3rd Qu.:49.10        3rd Qu.:4.000            3rd Qu.:40.4          
#>  Max.   :70.00        Max.   :8.000            Max.   :71.0          
#>  NA's   :8440         NA's   :8440             NA's   :8440          
#>   vanhanen_pmm   wahman_teorell_hadenius
#>  Min.   :1.000   Min.   :1.000          
#>  1st Qu.:1.000   1st Qu.:1.000          
#>  Median :2.000   Median :1.000          
#>  Mean   :2.939   Mean   :1.424          
#>  3rd Qu.:5.000   3rd Qu.:2.000          
#>  Max.   :8.000   Max.   :2.000          
#>  NA's   :15210   NA's   :17482