Prepares selected indexes in the democracy dataset for use in a UD model

Prepares selected indexes in the democracy dataset for use in a UD model

prepare_democracy(..., identifiers = c("extended_country_name", "year", "GWn",
  "cown", "in_GW_system"))

Arguments

...
a set of measures of democracy in the democracy dataset (either bare column names, or strings). Can be a dplyr::select expression (but dplyr must be loaded for this to work).

identifiers
A set of identifiers to return with the data. Defaults to c("extended_country_name", "year", "GWn", "cown", "in_GW_system"). Must exist in the democracy dataset.

Value

A dataset ready for use with democracy_model

Examples

prepare_democracy(dplyr::matches("pmm"))

#> # A tibble: 9,137 x 17
#>    extended_country_name  year   GWn  cown in_GW_system pmm_arat pmm_blm
#>                    <chr> <dbl> <dbl> <int>        <lgl>    <dbl>   <dbl>
#>  1           Afghanistan  1946   700   700         TRUE       NA      NA
#>  2           Afghanistan  1947   700   700         TRUE       NA      NA
#>  3           Afghanistan  1948   700   700         TRUE        2      NA
#>  4           Afghanistan  1949   700   700         TRUE        2      NA
#>  5           Afghanistan  1950   700   700         TRUE        2      NA
#>  6           Afghanistan  1951   700   700         TRUE        2      NA
#>  7           Afghanistan  1952   700   700         TRUE        2      NA
#>  8           Afghanistan  1953   700   700         TRUE        2      NA
#>  9           Afghanistan  1954   700   700         TRUE        2      NA
#> 10           Afghanistan  1955   700   700         TRUE        2      NA
#> # ... with 9,127 more rows, and 10 more variables: pmm_bollen <dbl>,
#> #   pmm_fh <dbl>, pmm_hadenius <dbl>, pmm_mainwaring <dbl>, pmm_munck <dbl>,
#> #   pmm_pacl <dbl>, pmm_polity <dbl>, pmm_polyarchy <dbl>, pmm_prc <dbl>,
#> #   pmm_vanhanen <dbl>